feat: enhance vet cli with file pattern and stdin features
